Update functions.php
As Shea pointed out on WPSE, it is more future-proof to use get_user_option():
If something changes in the MP6 plugin or it gets moved into core, as long as the pre_get_user_option filter is still in place (which it would be, to maintain backwards compat), the code will still work. Nobody’s going to fake an active plugin when MP6 is moved to core.